+++ ranger-r-package/ranger/R/ranger.R
@@ -33,12 +33,14 @@
 ##'
 ##' The tree type is determined by the type of the dependent variable.
 ##' For factors classification trees are grown, for numeric values regression trees and for survival objects survival trees.
-##' The Gini index is used as splitting rule for classification and the estimated response variances for regression.
+##' The Gini index is used as splitting rule for classification.
+##' For regression, the estimated response variances or maximally selected rank statistics (Wright et al. 2016) can be used.
 ##' For Survival the log-rank test, a C-index based splitting rule (Schmid et al. 2015) and maximally selected rank statistics (Wright et al. 2016) are available.
 ##'
 ##' With the \code{probability} option and factor dependent variable a probability forest is grown.
-##' Here, the estimated response variances are used for splitting, as in regression forests.
+##' Here, the node impurity is used for splitting, as in classification forests.
 ##' Predictions are class probabilities for each sample.
+##' In contrast to other implementations, each tree returns a probability estimate and these estimates are averaged for the forest probability estimate.
 ##' For details see Malley et al. (2012).
 ##'
 ##' Note that for classification and regression nodes with size smaller than min.node.size can occur, like in original Random Forest.
@@ -47,6 +49,13 @@
 ##' In \code{split.select.weights} variables weighted with 0 are never selected and variables with 1 are always selected. 
 ##' Weights do not need to sum up to 1, they will be normalized later. 
 ##' The usage of \code{split.select.weights} can increase the computation times for large forests.
+##'
+##' Unordered factor covariates can be handled in 3 different ways by using \code{respect.unordered.factors}: 
+##' For 'ignore' all factors are regarded ordered, for 'partition' all possible 2-partitions are considered for splitting. 
+##' For 'order' and 2-class classification the factor levels are ordered by their proportion falling in the second class, for regression by their mean response, as described in Hastie et al. (2009), chapter 9.2.4.
+##' For multiclass classification and survival outcomes, 'order' is experimental and should be used with care.
+##' The use of 'order' is recommended for 2-class classification and regression, as it computationally fast and can handle an unlimited number of factor levels. 
+##' Note that the factors are only reordered once and not again in each split. 
 ##'
 ##' For a large number of variables and data frame as input data the formula interface can be slow or impossible to use.
 ##' Alternatively dependent.variable.name (and status.variable.name for survival) can be used.
@@ -78,12 +87,12 @@
 ##' @param replace Sample with replacement. 
 ##' @param sample.fraction Fraction of observations to sample. Default is 1 for sampling with replacement and 0.632 for sampling without replacement. 
 ##' @param case.weights Weights for sampling of training observations. Observations with larger weights will be selected with higher probability in the bootstrap (or subsampled) samples for the trees.
-##' @param splitrule Splitting rule, survival only. The splitting rule can be chosen of "logrank", "C" and "maxstat" with default "logrank". 
+##' @param splitrule Splitting rule, regressiond and survival only. For regression one of "variance" or "maxstat" with default "variance". For survival "logrank", "C" or "maxstat" with default "logrank". 
 ##' @param alpha For "maxstat" splitrule: Significance threshold to allow splitting.
 ##' @param minprop For "maxstat" splitrule: Lower quantile of covariate distribtuion to be considered for splitting.
 ##' @param split.select.weights Numeric vector with weights between 0 and 1, representing the probability to select variables for splitting. Alternatively, a list of size num.trees, containing split select weight vectors for each tree can be used.  
 ##' @param always.split.variables Character vector with variable names to be always tried for splitting.
-##' @param respect.unordered.factors Regard unordered factor covariates as unordered categorical variables. If \code{FALSE}, all factors are regarded ordered. 
+##' @param respect.unordered.factors Handling of unordered factor covariates. One of 'ignore', 'order' and 'partition' with default 'ignore'. Alternatively TRUE (='order') or FALSE (='ignore') can be used. See below for details. 
 ##' @param scale.permutation.importance Scale permutation importance by standard error as in (Breiman 2001). Only applicable if permutation variable importance mode selected.
 ##' @param keep.inbag Save how often observations are in-bag in each tree. 
 ##' @param holdout Hold-out mode. Hold-out all samples with case weight 0 and use these for variable importance and prediction error.
@@ -161,6 +170,7 @@
 ##'   \item Breiman, L. (2001). Random forests. Mach Learn, 45(1), 5-32. 
 ##'   \item Ishwaran, H., Kogalur, U. B., Blackstone, E. H., & Lauer, M. S. (2008). Random survival forests. Ann Appl Stat, 841-860. 
 ##'   \item Malley, J. D., Kruppa, J., Dasgupta, A., Malley, K. G., & Ziegler, A. (2012). Probability machines: consistent probability estimation using nonparametric learning machines. Methods Inf Med, 51(1), 74.
+##'   \item Hastie, T., Tibshirani, R., Friedman, J. (2009). The Elements of Statistical Learning. Springer, New York. 2nd edition.
 ##'   }
 ##' @seealso \code{\link{predict.ranger}}
 ##' @useDynLib ranger
@@ -175,7 +185,7 @@ ranger <- function(formula = NULL, data = NULL, num.trees = 500, mtry = NULL,
                    case.weights = NULL, 
                    splitrule = NULL, alpha = 0.5, minprop = 0.1,
                    split.select.weights = NULL, always.split.variables = NULL,
-                   respect.unordered.factors = FALSE,
+                   respect.unordered.factors = "ignore",
                    scale.permutation.importance = FALSE,
                    keep.inbag = FALSE, holdout = FALSE,
                    num.threads = NULL, save.memory = FALSE,
@@ -255,24 +265,64 @@ ranger <- function(formula = NULL, data = NULL, num.trees = 500, mtry = NULL,
                                                           colnames(data.selected) != status.variable.name]
   }
   
-  ## Recode characters as factors
-  if (!is.matrix(data.selected)) {
-    char.columns <- sapply(data.selected, is.character)
-    data.selected[char.columns] <- lapply(data.selected[char.columns], factor)
+  ## Old version of if respect.unordered.factors
+  if (respect.unordered.factors == TRUE) {
+    respect.unordered.factors <- "order"
+  } else if (respect.unordered.factors == FALSE) {
+    respect.unordered.factors <- "ignore"
   }
   
-  ## Input data and variable names
+  ## Recode characters as factors and recode factors if 'order' mode
+  if (!is.matrix(data.selected)) {
+    character.idx <- sapply(data.selected, is.character)
+    
+    if (respect.unordered.factors == "order") {
+      ## Recode characters and unordered factors
+      names.selected <- names(data.selected)
+      ordered.idx <- sapply(data.selected, is.ordered)
+      factor.idx <- sapply(data.selected, is.factor)
+      independent.idx <- names.selected != dependent.variable.name & 
+        names.selected != status.variable.name & 
+        names.selected != paste0("Surv(", dependent.variable.name, ", ", status.variable.name, ")")
+      recode.idx <- independent.idx & (character.idx | (factor.idx & !ordered.idx))
+
+      ## Numeric response
+      if (is.factor(response)) {
+        num.response <- as.numeric(response)
+      } else if (!is.null(dim(response))) {
+        num.response <- response[, 1]
+      } else {
+        num.response <- response
+      }
+
+      ## Recode each column
+      data.selected[recode.idx] <- lapply(data.selected[recode.idx], function(x) {
+        ## Order factor levels
+        means <- aggregate(num.response~x, FUN=mean)
+        levels.ordered <- means$x[order(means$num.response)]
+        
+        ## Return reordered factor
+        factor(x, levels = levels.ordered)
+      })
+      
+      ## Save levels
+      covariate.levels <- lapply(data.selected[independent.idx], levels)
+    } else {
+      ## Recode characters only
+      data.selected[character.idx] <- lapply(data.selected[character.idx], factor)
+    }
+  }
+  
+  ## Input data and variable names, create final data matrix
   if (!is.null(formula) & treetype == 5) {
-    data.final <- cbind(response[, 1], response[, 2],
-                        data.selected[-1])
+    data.final <- data.matrix(cbind(response[, 1], response[, 2],
+                              data.selected[-1]))
     colnames(data.final) <- c(dependent.variable.name, status.variable.name,
                               independent.variable.names)
-  } else {
+  } else if (is.matrix(data.selected)) {
     data.final <- data.selected
-  }
-  if (!is.matrix(data.selected)) {
-    ## Create matrix
-    data.final <- data.matrix(data.final)
+  } else {
+    data.final <- data.matrix(data.selected)
   }
   variable.names <- colnames(data.final)
   
@@ -398,16 +448,42 @@ ranger <- function(formula = NULL, data = NULL, num.trees = 500, mtry = NULL,
   
   ## Splitting rule
   if (is.null(splitrule)) {
-    splitrule <- "logrank"
+    if (treetype == 5) {
+      splitrule <- "logrank"
+    } else if (treetype == 3) {
+      splitrule <- "variance"
+    }
     splitrule.num <- 1
-  } else if (treetype == 5 & splitrule == "logrank") {
-    splitrule.num <- 1
-  } else if (treetype == 5 & (splitrule == "auc" | splitrule == "C")) {
-    splitrule.num <- 2
-  } else if (treetype == 5 & (splitrule == "auc_ignore_ties" | splitrule == "C_ignore_ties")) {
-    splitrule.num <- 3
-  } else if (treetype == 5 & splitrule == "maxstat") {
-    splitrule.num <- 4
+  } else if (splitrule == "logrank") {
+    if (treetype == 5) {
+      splitrule.num <- 1
+    } else {
+      stop("Error: logrank splitrule applicable to survival data only.")
+    }
+  } else if (splitrule == "variance") {
+    if (treetype == 3) {
+      splitrule.num <- 2
+    } else {
+      stop("Error: variance splitrule applicable to regression data only.")
+    }
+  } else if (splitrule == "auc" | splitrule == "C") {
+    if (treetype == 5) {
+      splitrule.num <- 2
+    } else {
+      stop("Error: C index splitrule applicable to survival data only.")
+    }
+  } else if (splitrule == "auc_ignore_ties" | splitrule == "C_ignore_ties") {
+    if (treetype == 5) {
+      splitrule.num <- 3
+    } else {
+      stop("Error: C index splitrule applicable to survival data only.")
+    }
+  } else if (splitrule == "maxstat") {
+    if (treetype == 5 | treetype == 3) {
+      splitrule.num <- 4
+    } else {
+      stop("Error: maxstat splitrule applicable to regression or survival data only.")
+    }
   } else {
     stop("Error: Unknown splitrule.")
   }
@@ -421,7 +497,7 @@ ranger <- function(formula = NULL, data = NULL, num.trees = 500, mtry = NULL,
   }
 
   ## Unordered factors  
-  if (respect.unordered.factors) {
+  if (respect.unordered.factors == "partition") {
     names.selected <- names(data.selected)
     ordered.idx <- sapply(data.selected, is.ordered)
     factor.idx <- sapply(data.selected, is.factor)
@@ -435,17 +511,42 @@ ranger <- function(formula = NULL, data = NULL, num.trees = 500, mtry = NULL,
       max.level.count <- 8*.Machine$sizeof.pointer - 1
       if (max(num.levels) > max.level.count) {
         stop(paste("Too many levels in unordered categorical variable ", unordered.factor.variables[which.max(num.levels)], 
-                   ". Only ", max.level.count, " levels allowed on this system. Consider ordering this factor.", sep = ""))
+                   ". Only ", max.level.count, " levels allowed on this system. Consider using the 'order' option.", sep = ""))
       } 
     } else {
       unordered.factor.variables <- c("0", "0")
       use.unordered.factor.variables <- FALSE
     } 
-  } else {
+  } else if (respect.unordered.factors == "ignore" | respect.unordered.factors == "order") {
+    ## Ordering for "order" is handled above
     unordered.factor.variables <- c("0", "0")
     use.unordered.factor.variables <- FALSE
+  } else {
+    stop("Error: Invalid value for respect.unordered.factors, please use 'order', 'partition' or 'ignore'.")
+  }
+
+  ## Unordered maxstat splitting not possible
+  if (use.unordered.factor.variables & !is.null(splitrule)) {
+    if (splitrule == "maxstat") {
+      stop("Error: Unordered factor splitting not implemented for 'maxstat' splitting rule.")
+    } else if (splitrule %in% c("C", "auc", "C_ignore_ties", "auc_ignore_ties")) {
+      stop("Error: Unordered factor splitting not implemented for 'C' splitting rule.")
+    }
   }
   
+  ## Warning for experimental 'order' splitting 
+  if (respect.unordered.factors == "order") {
+    if (treetype == 5) {
+      warning("Warning: The 'order' mode for unordered factor handling for survival outcomes is experimental.")
+    } else if (treetype == 1 | treetype == 9) {
+      if (nlevels(response) > 2) {
+        warning("Warning: The 'order' mode for unordered factor handling for multiclass classification is experimental.")
+      }
+    } else if (treetype == 3 & splitrule == "maxstat") {
+      warning("Warning: The 'order' mode for unordered factor handling with the 'maxstat' splitrule is experimental.")
+    }
+  }
+
   ## Prediction mode always false. Use predict.ranger() method.
   prediction.mode <- FALSE
   predict.all <- FALSE
@@ -485,9 +586,15 @@ ranger <- function(formula = NULL, data = NULL, num.trees = 500, mtry = NULL,
     result$chf <- result$predictions
     result$predictions <- NULL
     result$survival <- exp(-result$chf)
-    result$splitrule <- splitrule
   } else if (treetype == 9 & !is.matrix(data)) {
-    colnames(result$predictions) <- levels(droplevels(response))
+    ## Set colnames and sort by levels
+    colnames(result$predictions) <- unique(response)
+    result$predictions <- result$predictions[, levels(droplevels(response))]
+  }
+  
+  ## Splitrule
+  if (treetype == 3 | treetype == 5) {
+    result$splitrule <- splitrule
   }
   
   ## Set treetype
@@ -513,6 +620,11 @@ ranger <- function(formula = NULL, data = NULL, num.trees = 500, mtry = NULL,
     result$forest$independent.variable.names <- independent.variable.names
     result$forest$treetype <- result$treetype
     class(result$forest) <- "ranger.forest"
+    
+    ## In 'ordered' mode, save covariate levels
+    if (respect.unordered.factors == "order" & !is.matrix(data)) {
+      result$forest$covariate.levels <- covariate.levels
+    }
   }
   
   class(result) <- "ranger"