+++ NEWS.md
@@ -1,3 +1,21 @@
+##### Version 0.5.2
+* Use unadjusted p-value for 2 categories in maxstat splitting
+
+##### Version 0.5.1
+* Bug fixes
+
+##### Version 0.5.0
+* New CRAN version.
+
+##### Version 0.4.7
+* Add splitting by maximally selected rank statistics for regression forests
+
+##### Version 0.4.6
+* Bug fixes
+
+##### Version 0.4.5
+* Use faster method for unordered factor splitting
+
 ##### Version 0.4.4
 * Add p-values for variable importance
 * Bug fixes